Genixer_Shikra/requirements.txt
datasets==2.19.0
einops==0.8.0
flash_attn==2.5.8
matplotlib==3.8.4
mmengine==0.10.4
numpy==1.26.4
opencv_python==4.9.0.80
peft==0.10.0
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.37.2

nltk
tqdm
accelerate
SentencePiece
gradio
fastapi
uvicorn
bitsandbytes